+++ backend/src/app/controllers/authController.js
@@ -2,9 +2,9 @@ const express = require('express');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 
-const authConfig = require('../config/auth.json');
+const authConfig = require('../../config/auth.json');
 
-const User = require('../models/User');
+const User = require('../../models/User');
 
 const router = express.Router();
