+++ backend/src/app/middlewares/auth.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken');
-const authConfig = require('../config/auth.json');
+const authConfig = require('../../config/auth.json');
 
 /**
  * Middleware for JWT validation
